fix: move unit tests in codex-rs/core/src/config/mod.rs into their own file
